pyproject.toml
[project]
name = "multi-agent-customer-support"
version = "0.1.0"
description = "Collaborative Multi-Agent System for Customer Support Automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langgraph>=0.0.20",
    "chromadb>=0.4.22",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scipy>=1.12.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "datasets>=2.14.0",
    "httpx>=0.26.0",
    "google-generativeai>=0.8.0",
]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]
